+++ uv_copy_and_paste_uv/__init__.py
@@ -21,7 +21,7 @@
 __author__ = "Nutti <nutti.metro@gmail.com>"
 __status__ = "production"
 __version__ = "3.0"
-__date__ = "X XXXX 2015"
+__date__ = "15 Jun 2015"
 
 bl_info = {
     "name": "Copy and Paste UV",
