+++ uv_copy_and_paste_uv/cpuv_common.py
@@ -25,7 +25,7 @@ from collections import namedtuple
 __author__ = "Nutti <nutti.metro@gmail.com>"
 __status__ = "production"
 __version__ = "3.0"
-__date__ = "X XXXX 2015"
+__date__ = "15 Jun 2015"
 
 SelectedFaceInfo = namedtuple('SelectedFaceInfo', 'normal indices center')
